ch4

+++ src/matrix/invert_matrix_utils.c
@@ -24,7 +24,7 @@ t_matrix *create_cofactor_matrix(t_matrix m)
     return (result);
 }
 
-t_matrix *divide_matrix_by(t_matrix m, int num)
+t_matrix *divide_matrix_by(t_matrix m, float num)
 {
     t_matrix *result;
     int i;
